requirements.txt
flask
pytesseract
Pillow
pymupdf
gspread
oauth2client